r8 - testing the cmod_post
- OOP, data modeling and encapsulation works fine.
- minor glitches resolved.
- JSON files executes. In order for the json files to be saved. Their
directories must exist beforehand.
- It is recommended to start off with new database instead of stacking
sample_db to the existing database. Linking to the user_table of another
db is not difficult. The reasoning is that, there is submission loop for
F5 keypress or back&re-submit, no page redirection yet. This must be
done with jQuery Ajax submit for a full solution.